requirements.txt
sounddevice
mlx-whisper
rumps
pynput
pyperclip
numpy
pyobjc-framework-EventKit
pyobjc-framework-Cocoa
